### 2021-07-29 #3548 @whedon
[PRE REVIEW]: bmm: Bayesian Map-matching
**Submitting author:** @SamDuffield (<a href="http://orcid.org/0000-0002-8656-8734">Samuel Duffield</a>)
**Repository:** <a href="https://github.com/SamDuffield/bmm" target ="_blank">https://github.com/SamDuffield/bmm</a>
**Version:** v1.1
