### 2021-07-29 #3548 @kyleniemeyer
Hi @SamDuffield, thanks for your interest in JOSS. We don't have any editors in this area with availability right now, so your submission will be placed in our waitlist until someone opens up.

In the meantime, your article is *extremely* short. Although JOSS articles are indeed short, we typically expect 250-1000 words—I would not usually count the Acknowledgments section, which is around the same length as the main content of your article.

Actually, it looks like some of what is in that section should actually belong in the main part of the article, since it is describing dependencies and things that your software is built on. Could you perhaps move that content (and anything else along those lines) to the main text? I think you could also explain how the software works in a bit more detail.
