+++ paper.md
@@ -17,11 +17,23 @@ bibliography: paper.bib
 
 # Summary
 
-`bmm` provides probabilistic map-matching with uncertainty quantification, built on top of `OSMnx` [@Boeing2017].
+`bmm` is a Python package providing probabilistic map-matching with uncertainty quantification.
 Map-matching is the task of converting a polyline (series of noisy location observations - e.g. GPS data)
-and a graph (collection of edges and nodes) into a continuous route trajectory restricted to the graph.
-`bmm` uses Bayesian particle smoothing methods to produce a collection of particles, each of which representing a
-continuous, plausible route along edges in the graph.
+and a graph (collection of edges and nodes) into a continuous route trajectory restricted to the graph. Here a
+continuous route is represented by series of connected edges as well as positions along said edges at
+observation time. `bmm` uses Bayesian particle smoothing methods to produce a collection of particles, each of which
+representing a continuous, plausible route along edges in the graph.
+
+`bmm` is built on top of `OSMnx` [@Boeing2017] - a python package assisting with the retrieval and processing
+of OpenStreetMap data [@OpenStreetMap]. Although, `bmm` is applicable to be used on any suitably
+labelled NetworkX graph [@Hagberg2008].
+
+In addition, `bmm` utilises `numpy` [@Harris2020] and `numba` [@Lam2015] for fast scientific calculations,
+`pandas` [@Reback2020] and `geopandas` [@Jordahl2020] for spatial data storage and manipulation
+as well as `matplotlib` [@Hunter2007] for visualisation.
+
+Documentation for `bmm` can be found at [bmm.readthedocs.io](https://bmm.readthedocs.io/en/latest/).
+
 
 # Statement of need
 
@@ -40,16 +52,59 @@ uncertainty in both the route taken and the positions at observation times. Addi
 support for both offline and online computation.
 
 
+# Core Functionality
+
+`bmm` can be used to convert a polyline (ordered series of GPS coordinates) into a collection of possible routes
+along edges within a graph.
+
+We assume that the graph is stored as a NetworkX [@Hagberg2008] object (which can easily be
+achieved for a given region using `OSMnx` [@Boeing2017]) and that the polyline is stored as an array or list of
+two-dimensional coordinates in the same coordinate system as the graph. A common choice for coordinate system
+is UTM (Universal Transverse Mercator) which as a square coordinate system (with unit metres) is less
+cumbersome than the spherical longitude-latitude coordinates system (with unit degrees). `bmm` can convert
+longitude-latitude to UTM using the `bmm.long_lat_to_utm` function.
+
+Given a suitable graph and polyline `bmm` can be easily used to map-match
+```python
+matched_particles = bmm.offline_map_match(graph, polyline=polyline_utm, n_samps=100, timestamps=15)
+```
+Here the `n_samps` parameter represents the number of particles/trajectories to output and `timestamps` is the
+number of seconds between polyline observations - this can be a float if all observation times are equally spaced,
+an array of length one less than that of the polyline representing the unequal times between observations or an 
+array of length equal to the polyline representing UNIX timestamps for the observation times.
+
+The output of `bmm.offline_map_match` is a `bmm.MMParticles` object that contains a `particles` attributes listing
+the possible trajectories the algorithm has managed to fit to the polyline - full details can be found at
+[bmm.readthedocs.io](https://bmm.readthedocs.io/en/latest/).
+
+`bmm` can also map-match data that arrives in an online or sequential manner. Initiate a `bmm.MMParticles`
+with the first observation
+```python
[…]
