+++ paper.md
@@ -96,9 +96,7 @@ Once a polyline has been succesfully map-matched, it can be visualised using `bm
 ```python
 bmm.plot(graph, particles=matched_particles, polyline=polyline_utm)
 ```
-
 ![](https://github.com/SamDuffield/bmm/blob/master/simulations/porto/test_route.png)
-Polyline matched using `bmm`.
 
 
 
