### 2021-07-30 #3548 @SamDuffield
Hi @kyleniemeyer! Thanks for your comments - I have moved the dependencies into the summary and extended the paper to properly describe the core functionality of the package.
