+++ paper.bib
@@ -8,8 +8,8 @@ volume = {65},
 pages = {126-139},
 year = {2017},
 issn = {0198-9715},
-doi = {https://doi.org/10.1016/j.compenvurbsys.2017.05.004},
-url = {https://www.sciencedirect.com/science/article/pii/S0198971516303970},
+doi = {doi.org/10.1016/j.compenvurbsys.2017.05.004},
+url = {www.sciencedirect.com/science/article/pii/S0198971516303970},
 author = {Geoff Boeing},
 keywords = {Street network, Urban form, Urban design, Transportation, Resilience, OpenStreetMap, GIS, Complex networks,
 Visualization, Python},
