### 2021-08-17 #3548 @SamDuffield
**Abstract**

`bmm`: Bayesian Map-matching

Map-matching is defined as the task of converting an ordered series of noisy, unconstrained GPS coordinates into a series of connected edges that represent the route taken constrained to a predefined road network. `bmm` represents a Python package for map-matching in dense, urban networks that combines a Bayesian state-space model with particle smoothing (or sequential Monte Carlo) to provide map-matching inference with uncertainty quantification in both offline and online settings.
