### 2021-09-21 SamDuffield/bayesian-traffic#2 @matt-graham
`ResourceWarning` on running `bmm/src/tests/test_smc.py` due to not closing files
*Submitting as part of JOSS review openjournals/joss-reviews/issues/3651*

The `test_data` attribute set in the test set up routine

https://github.com/SamDuffield/bmm/blob/df10de7edc650f542c85039bdf6ca447faa12ea9/bmm/src/tests/test_smc.py#L63

is set by calling `load_test_data`

https://github.com/SamDuffield/bmm/blob/df10de7edc650f542c85039bdf6ca447faa12ea9/bmm/src/tests/test_smc.py#L30-L43

which itself calls `read_data`

https://github.com/SamDuffield/bmm/blob/df10de7edc650f542c85039bdf6ca447faa12ea9/bmm/src/tests/test_smc.py#L21-L27

which due to the `chunksize` argument passed to  `pandas.read_csv` returns as a Pandas `TextFileReader`.

The `TextFileReader` has a `close` method which should be called when no longer needed to release the associated file resource - this could be done for example by defining a corresponding `tearDown` method in the `testWithGraphAndData` test case class which calls the `close` method. At the moment I get a series of `ResourceWarning: unclosed file <_io.TextIOWrapper ...>` warnings due to the files not being closed when trying to run the test script.
