### 2021-09-21 SamDuffield/bmm#3 @matt-graham
Tests in `bmm/src/tests/test_smc.py` fail with `AttributeError`
*Submitting as part of JOSS review openjournals/joss-reviews/issues/3651*

Running

```
python bmm/src/tests/test_smc.py
```

results in two test failures in `TestUpdateParticlesPF.test_update` and `TestUpdateParticlesBSi.test_update`, in both cases due to similar `AttributeError` instances

```
Traceback (most recent call last):
  File "bmm/bmm/src/tests/test_smc.py", line 102, in test_update
    updated_particles = smc.update_particles_flbs(self.graph,
  File "bmm/bmm/src/inference/smc.py", line 250, in update_particles_flbs
    out_particles.time_intervals = np.append(out_particles.time_intervals, time_interval)
AttributeError: 'MMParticles' object has no attribute 'time_intervals'
```

```
Traceback (most recent call last):
  File "/bmm/bmm/src/tests/test_smc.py", line 102, in test_update
    updated_particles = smc.update_particles_flbs(self.graph,
  File "bmm/bmm/src/inference/smc.py", line 250, in update_particles_flbs
    out_particles.time_intervals = np.append(out_particles.time_intervals, time_interval)
AttributeError: 'MMParticles' object has no attribute 'time_intervals'
```
