### 2021-09-21 SamDuffield/bmm#4 @matt-graham
Installing version 1.1 from PyPI fails due to missing `requirements.txt`
*Submitting as part of JOSS review openjournals/joss-reviews/issues/3651*

On trying to run `pip install bmm` in a fresh Python 3.9 environment I get an error

```
    ...
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "setup.py", line 10, in <module>
        with open('requirements.txt') as f:
    FileNotFoundError: [Errno 2] No such file or directory: 'requirements.txt'
```

Inspecting the [`bmm-1.1.tar.gz` archive hosted by PyPI](https://files.pythonhosted.org/packages/49/60/45621c9c0b5c2a47d6ad19f46b3e4d49e488f109c1dd92ed47ca0879e1b9/bmm-1.1.tar.gz) it looks the `requirments.txt` file is not successfully being included in the uploaded package resulting in this error when installing with `pip` (with `pip` falling back to installing version 1.0 for which this error does not seem to occur).

I think you need to explicitly include the `requirements.txt` in the package by [creating a `MANIFEST.in` file](https://packaging.python.org/guides/using-manifest-in/).
