### 2021-09-21 SamDuffield/bmm#5 @matt-graham
Tests currently hard to find
*Submitting as part of JOSS review openjournals/joss-reviews/issues/3651*

The tests currently live in `bmm/src/tests` and aren't signposted anywhere from the main README file or the documentation as far as I can see. As they are not part of the `bmm` package given the lack of an `__init__.py` file in the `bmm/src/tests` I would say it makes more sense to place them outside the package directory for example in a `tests` directory at the root of the repository.
