+++ simulations/porto/utils.py
@@ -127,243 +127,8 @@ def interval_tv_dists(particles_one,
     return tv_each_time
 
 
-def total_variation_edges(edges_one,
-                          edges_two,
-                          round_alpha=None):
-    n1 = len(edges_one)
-    n2 = len(edges_two)
-
-    if edges_one.shape[-1] == 4 and round_alpha is not None:
-        edges_one[:, -1] = np.round(edges_one[:, -1], round_alpha)
-        edges_two[:, -1] = np.round(edges_two[:, -1], round_alpha)
-
-    all_edges = np.concatenate([edges_one, edges_two])
-    all_edges = np.unique(all_edges, axis=0)
-
-    axes = tuple(i for i in range(1, all_edges.ndim))
-
-    tv = 0.
-    for edge in all_edges:
-        p_1 = np.sum(np.all(np.abs(edges_one - edge) < 1e-3, axis=axes)) / n1
-        p_2 = np.sum(np.all(np.abs(edges_two - edge) < 1e-3, axis=axes)) / n2
-        tv = tv + np.abs(p_1 - p_2)
-
-    return tv / 2
-
-
-def each_edge_route_total_variation(particles_one,
-                                    particles_two,
-                                    observation_times,
-                                    include_alpha=False,
-                                    round_alpha=None):
-    m = observation_times.size
-    tv_each_time = np.zeros(m)
-
-    alpha_extend = include_alpha * 1
-
-    for i in range(m):
-        if i == 0:
-            p1_first_edges = np.array([p[:1, 1:(4 + alpha_extend)] for p in particles_one])
-            p2_first_edges = np.array([p[:1, 1:(4 + alpha_extend)] for p in particles_two])
-
-            tv_each_time[i] = total_variation_edges(p1_first_edges, p2_first_edges, round_alpha)
-        else:
-            prev_time = observation_times[i - 1]
-            current_time = observation_times[i]
-
-            p1_edges = np.zeros((len(particles_one), 1, (3 + alpha_extend)))
-            for j, p1 in enumerate(particles_one):
-                prev_ind = np.where(p1[:, 0] == prev_time)[0][0]
-                curr_ind = np.where(p1[:, 0] == current_time)[0][0]
-
-                p1_ed = p1[(prev_ind + 1):(curr_ind + 1), 1:(4 + alpha_extend)].copy()
-                p1_ed_len = len(p1_ed)
-                if p1_ed_len > p1_edges.shape[1]:
-                    p1_edges = np.append(p1_edges, np.zeros((p1_edges.shape[0],
-                                                             p1_ed_len - p1_edges.shape[1],
-                                                             (3 + alpha_extend))),
-                                         axis=1)
-
-                p1_edges[j, :p1_ed_len] = p1_ed
-
-            p2_edges = np.zeros((len(particles_two), p1_edges.shape[1], (3 + alpha_extend)))
-            for j, p2 in enumerate(particles_two):
-                prev_ind = np.where(p2[:, 0] == prev_time)[0][0]
-                curr_ind = np.where(p2[:, 0] == current_time)[0][0]
-
-                p2_ed = p2[(prev_ind + 1):(curr_ind + 1), 1:(4 + alpha_extend)].copy()
-                p2_ed_len = len(p2_ed)
-                if p2_ed_len > p2_edges.shape[1]:
-                    p2_edges = np.append(p2_edges, np.zeros((p2_edges.shape[0],
[…]
