### 2021-10-06 SamDuffield/bmm#2 @SamDuffield
Thanks for pointing this out. I have replaced the `chunksize` argument with `nrows` so that `read_data` always returns a `pandas.DataFrame` instead
