### 2021-10-06 SamDuffield/bmm#3 @SamDuffield
Thanks - fixed to create an empty `time_intervals` if it hasn't been initiated yet (as is common at the start of a trip)
