### 2021-10-06 SamDuffield/bmm#4 @SamDuffield
`MANIFEST.in` is in!
