### 2021-10-06 SamDuffield/bmm#5 @SamDuffield
Yes good idea! Have moved them
