### 2021-10-08 SamDuffield/bmm#4 @matt-graham
Thanks for fixing this - confirming I can now successfully install using `pip`.
