### 2021-10-08 SamDuffield/bmm#6 @matt-graham
Possible typo in docstring
*Submitting as part of JOSS review openjournals/joss-reviews/issues/3651*

In the following

https://github.com/SamDuffield/bmm/blob/216edb8f1ebe04fbf9745b1d9988ee1ccaaba467/bmm/src/inference/particles.py#L130-L134

it seem like the `shape` specified in the final line should perhaps be `(_, 8)` instead as this would be consistent with the earlier definition in

https://github.com/SamDuffield/bmm/blob/216edb8f1ebe04fbf9745b1d9988ee1ccaaba467/bmm/src/inference/particles.py#L21-L30

It would also help to explicitly explain what the `_`  dimension corresponds to in both these cases.
