### 2021-10-08 SamDuffield/bayesian-traffic#7 @matt-graham
Sanity check graph takes a long time to load
*Submitting as part of JOSS review openjournals/joss-reviews/issues/3651*

I really like the sanity check example in the documentation, however I think it may be worth using a different location than London to source the example graph as running

```Python
graph = ox.graph_from_place('London, UK')
```

takes around 8 minutes on my laptop and the lack of any progress updates from the `osmnx` package when executing `graph_from_place` made it a bit unclear initially if the interpreter had crashed. Swapping for a location that produces a smaller graph would still perform the same function of illustrating the inference algorithm is computing something sensible but without requiring so much patience from the user! If you prefer to keep the current example, I think it would then be worth at least putting a warning to expect the `osmnx` functions to take a long time to run.
