### 2021-10-08 SamDuffield/bmm#8 @matt-graham
Allow using fixing the initial random number generator state for reproducibility
*Submitting as part of JOSS review openjournals/joss-reviews/issues/3651*

Currently random variates are generated throughout the package by calls to the functions in `numpy.random` which [uses a global `RandomState` instance](https://numpy.org/doc/stable/reference/random/legacy.html#functions-in-numpy-random) which is automatically seeded with entropy from the local device. As the linked documentation notes this is discouraged both because it uses global state and uses the legacy `RandomState` rather than the more modern (and performant) `Generator`.

Adding an optional argument to functions which currently make calls to `numpy.random` corresponding to a `RandomState` to be used in place of `numpy.random` would make it simpler to ensure reproducibility of the outputs produced, for example when writing unit tests. Compatibility with the current interface could still be maintained by defaulting to `numpy.random` if this argument is not specified. While `numpy.random` does include a `seed` function which allows setting a seed for the global `RandomState` instance, even when fixing the seed the output from a function which uses this global state can differ from run to run if the other related code run before or afterwards which also uses the global state is changed.

Even better would be to use [the updated `Generator` class](https://numpy.org/doc/stable/reference/random/generator.html) however this would involve a bit more effort to update as the naming and interface of some methods has changed compared to `RandomState`.
