### 2021-10-08 SamDuffield/bmm#9 @matt-graham
Test coverage could be improved
*Submitting as part of JOSS review openjournals/joss-reviews/issues/3651*

The current tests only perform quite minimal checks with the code and coverage is sparse. While writing tests for correctness of implementations of Monte Carlo algorithms is challenging, I think there are lots of additional simple checks that could be added without having to delve too much in to the details of performing statistical hypothesis tests etc.

As some indicative examples of where test coverage could be improved:

  * Currently the tests in `tests/test_resampling.py` only check for cases where the particle weights correspond to all the probability mass being on one particle (and so the resampling step is deterministic). While this is definitely a useful edge case to test for, it would be useful to also have checks for more typical cases where the particle weights are non-degenerate. For example testing that the frequencies of particle indices sampled over a large number of replications are within some tolerance of the particle weights.
  * The tests in `tests/test_smc.py` currently only perform very minimal checks - for example that the number of particles is as expected. There are a lot of additional simple checks that could be performed here such as checking that the particle weights are finite and non-negative, that the updated particles have finite / non-`NaN` values and differ from the previous particles, that different observations lead to different sampled values for the particles.
  * For the function ` get_possible_routes` in `bmm/src/inference/proposal.py` it seems like it should be possible to come up with a test for correctness on a trivial graph / route for which its possible to manually enumerate all possible routes.
  * It seems that there are several cases where functions could be tested for consistency against each other in particular regimes - for example should `full_fixed_lag_stitch` and `rejection_fixed_lag_stitch` give (statistically) consistent outputs as `max_rejections` goes to infinity, and similarly for `full_backward_sample` and `rejection_backward_sample`?
  * `backwards_simulate` from `bmm/src/inference/backward.py` could be checked to ensure that the locations of the particles in the returned particle trajectories do all correspond to the inputted particles from the forward filtering, that we get the expected output when the the filtering weights are non-zero for only one particle at each observation time.
  * It seems like a simplified version of the sanity check example using a very simple graph could be written in to a test of some sorts.
