### 2021-10-11 fab9a1a
amend particles docstring
