### 2021-10-11 SamDuffield/bmm#7 @SamDuffield
As a large portion of the anticipated use cases is in urban environments (most likely major cities such as London) I think leaving the example as is but adding a precautionary comment about `osmnx` loading times is the most useful for the user. (Also the `sample_route` function works better for larger areas as it is less likely to terminate early on reaching the edge of the area.)

I have added said warnings to the docs!
