+++ CONTRIBUTING.md
@@ -1,0 +1,44 @@
+# Contributing to Transcriptase
+`bmm` would love contributions! Whether that's:
+
+- Reporting a bug
+- Discussing the current state of the code
+- Submitting a fix
+- Proposing new features
+- Becoming a maintainer
+
+## We Develop with GitHub
+`bmm` uses GitHub to host code, track issues and feature requests, as well as accepting pull requests.
+
+## We Use [GitHub Flow](https://guides.github.com/introduction/flow/index.html), So All Code Changes Happen Through Pull Requests
+Pull requests are the best way to propose changes to the codebase (we use [GitHub Flow](https://guides.github.com/introduction/flow/index.html)). We actively welcome your pull requests:
+
+1. Fork the repo and create your branch from `master`.
+2. If you've added code that should be tested, add tests.
+3. If you've changed APIs, update the documentation.
+4. Ensure the test suite passes.
+5. Make sure your code lints.
+6. Issue that pull request!
+
+## Any contributions you make will be under the MIT Software License
+In short, when you submit code changes, your submissions will be understood to be under the same [MIT License](https://github.com/SamDuffield/bmm/blob/master/LICENSE.txt) that covers the project. Feel free to contact the maintainers if that's a concern.
+
+## Report bugs using GitHub's [issues](https://github.com/SamDuffield/bmm/issues)
+We use GitHub issues to track public bugs. Report a bug by [opening a new issue](https://github.com/SamDuffield/bmm/issues/new); it's that easy!
+
+## Write bug reports with detail, background, and sample code
+**Great Bug Reports** tend to have:
+
+- A quick summary and/or background
+- Steps to reproduce
+- What you expected would happen
+- What actually happens
+- Notes (possibly including why you think this might be happening, or stuff you tried that didn't work)
+
+People *love* thorough bug reports. I'm not even kidding.
+
+## License
+By contributing, you agree that your contributions will be licensed under its [MIT License](https://github.com/SamDuffield/bmm/blob/master/LICENSE.txt).
+
+## References
+This document was adapted from [briandk](https://gist.github.com/briandk/3d2e8b3ec8daf5a27a62) and [Facebook's Draft](https://github.com/facebook/draft-js/blob/master/CONTRIBUTING.md).
