+++ README.md
@@ -3,6 +3,8 @@ Map-matching using particle smoothing methods.
 
 [Docs](https://bmm.readthedocs.io/en/latest/) and [methodology](https://arxiv.org/abs/2012.04602).
 
+Contributing guidelines can be found in the repo's `CONTRIBUTING.md` file.
+
 ## Install
 ```
 pip install bmm
