### 2021-10-12 5ebbaf0
add software links to paper
