### 2022-01-06 #3651 @georgiastuart
I am deeply sorry about the extreme delay. I have not been able to install this package and have only recently had time to really mess with it. It appears that the current BMM dependencies don't play nicely with my version of Ubuntu (21.10). @SamDuffield would it be possible for you to add a Dockerfile to the repo so I can build a container to test?
