### 2022-01-06 #3651 @georgiastuart
I was able to put together a quick Dockerfile. Continuing my review now.
