### 2022-02-04 #3651 @georgiastuart
Aside from the installation difficulties on Ubuntu 21.10 (which is more of a quirk of the OS and not, I think, something the author needs to address), BMM works as advertised. I found the documentation to be accessible and thorough and I was able to reproduce the examples. Pip installation works fine on Ubuntu 20.04 LTS and Rocky 8. I was not able to find another package that replicates the functionality of BMM, so I believe it represents a substantial contribution.

This is more of a "nice to have" feature, not a requirement for publication, but it was not clear to me how to save the result of the map matching (aside from, I guess, saving a pickle?). Some guidance on how to save results to disk would be very helpful.

The author updated the repository in response to all issues raised by @matt-graham . I think it's ready for publication.

Again, I apologize for the great delay.
