+++ paper.bib
@@ -16,13 +16,15 @@ Visualization, Python},
 }
 
 
-@misc{Duffield2020,
-      title={Online Particle Smoothing with Application to Map-matching},
-      author={Samuel Duffield and Sumeetpal S. Singh},
-      year={2020},
-      eprint={2012.04602},
-      archivePrefix={arXiv},
-      primaryClass={stat.ME}
+@ARTICLE{Duffield2020,
+  author={Duffield, Samuel and Singh, Sumeetpal S.},
+  journal={IEEE Transactions on Signal Processing}, 
+  title={Online Particle Smoothing With Application to Map-Matching}, 
+  year={2022},
+  volume={70},
+  number={},
+  pages={497-508},
+  doi={10.1109/TSP.2022.3141259}
 }
 
 
