### 2022-02-09 #3651 @taless474
The next step is to issue a new tagged release of the software (if changed) and archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), as explained [here](https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance).
Thank you, @SamDuffield, and thank you @matt-graham and @georgiastuart for all your effort.
