### 2022-02-10 #3651 @SamDuffield
> @SamDuffield LGTM! Would you please change the title in zanodo to match the title of your paper? Typically, these are the same for a JOSS-related paper. Note that you can manually change the metadata of the zenodo archive.

Ok should have matching title now!
